+++ app/Http/Controllers/LoginControllers/AffiliatesLoginController.php
@@ -21,7 +21,7 @@ class AffiliatesLoginController extends Controller
     }
     public function login(Request $request){
         if(Auth::guard('web')->attempt(['email'=>$request->email,'password'=>$request->password],$request->remember)){
-            return redirect()->intended('/administrator/maintenance/employees');
+            return redirect()->intended('/affiliates/maintenance/position');
         }
     }
     public function showRegister(){
@@ -36,6 +36,8 @@ class AffiliatesLoginController extends Controller
         $company->strCompanyName = $request->input('companyname');
         $company->strCompanyAddress = $request->input('address');
         $company->strCompanyEmail = $request->input('email');
+        $company->strCompanyContactPNum = $request->input('mobilenum');
+        $company->strCompanyContactTNum = $request->input('telnum');
         $company->save();
         // return response(['username'=>$request->input('username')]);
         $user = User::create([