+++ resources/views/Hauling/info.blade.php
@@ -1,0 +1,91 @@
+<!-- More Info Modal -->
+<div class="modal animated bounceIn fast" id="moreInfoModal" tabindex="-1" role="dialog" aria-labelledby="moreInfoModalLabel" aria-hidden="true">
+    <div class="modal-dialog modal-dialog-centered modal-lg" role="document">
+        <div class="modal-content">
+            <div class="modal-header">
+
+                <div class="modal-title" id="moreInfoModalLabel"><small>Job Order # 4</small>
+                    <h4>Consignee Name</h4></div>
+                <button type="button" class="mb-1 close modalClose" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body modalBody">
+                <div class="row mt-2">
+                    <div class="col-6">
+                        <ul class="list-inline">
+                            <li class="list-inline-item text-primary">
+                                <h6>Date of Transaction : </h6></li>
+                            <li class="list-inline-item">
+                                <h6>May 4, 2018</h6></li>
+                        </ul>
+                        <ul class="list-inline">
+                            <li class="list-inline-item text-primary">
+                                <h6>Estimated Time of Hauling : </h6></li>
+                            <li class="list-inline-item">
+                                <h6>0730 HRS</h6></li>
+                        </ul>
+                    </div>
+                    <div class="col-6">
+                        <ul class="list-inline">
+                            <li class="list-inline-item text-primary">
+                                <h6>Starting Location : </h6></li>
+                            <li class="list-inline-item">
+                                <h6>PUP</h6></li>
+                        </ul>
+                        <ul class="list-inline">
+                            <li class="list-inline-item text-primary">
+                                <h6>Destination : </h6></li>
+                            <li class="list-inline-item">
+                                <h6>Pureza</h6></li>
+                        </ul>
+                        <ul class="list-inline">
+                            <li class="list-inline-item text-primary">
+                                <h6>Goods to be delivered : </h6></li>
+                            <li class="list-inline-item">
+                                <h6>Very Good</h6></li>
+                        </ul>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-12">
+                        <table class="detailedTable text-center table table-striped fadeIn animated" style="width:100%">
+                            <thead class="bg-primary">
+                                <tr>
+                                    <th>Location</th>
+                                    <th>Time Updated</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                <tr>
+                                    <td>Pasig</td>
+                                    <td>24:00</td>
+                                </tr>
+                            </tbody>
+                        </table>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-12 mt-2 text-center">
+                        <div class="text-primary mb-2">
+                            <h4>Extra Details</h4></div>
+                        <div class="border-primary">
+                            <div class="mr-2 ml-2 mt-2 mb-2">
+                                <p>
+                                    When things go wrong, as they sometimes will, When the road you're trudging seems all uphill,
+                                    <br> When the funds are low and the debts are high,
+                                    <br> And you want to smile, but you have to sigh,
+                                    <br> When care is pressing you down a bit-
+                                    <br> Rest if you must, but don't you quit.
+                                    <br> Life is queer with its twists and turns,
+                                    <br>
+
+                                </p>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>