+++ resources/views/Hauling/scripts.blade.php
@@ -1,0 +1,2 @@
+    <script src="/others/stisla_admin_v1.0.0/dist/js/hauling/hauling.js"></script>
+    <script src="/others/stisla_admin_v1.0.0/dist/js/hauling/haulingStyles.js"></script>