+++ resources/views/Templates/newSidebar.blade.php
@@ -44,11 +44,10 @@
                                 <li id="goodsMenu"><a href="/administrator/maintenance/goods"><i class="fas fa-cube"></i>Goods</a></li>
                                 <li id="tugboatTypeMenu"><a href="/administrator/maintenance/tugboattype"><i class="ion ion-android-boat"></i>Tugboat Type</a></li>
                                 <li id="tugboatMenu"><a href="/administrator/maintenance/tugboat"><i class="ion ion-android-boat"></i>Tugboats</a></li>
-                                <li id="quotationsMenu"><a href="/administrator/maintenance/quotations"><i class="fas fa-life-ring"></i>Quotations</a></li>
                             @elseif(Auth::user()->enumUserType == 'Affiliates')
-                                <li id="positionMenu"><a href="/administrator/maintenance/position"><i class="fas fa-sitemap"></i>Position</a></li>
-                                <li id="employeesMenu"><a href="/administrator/maintenance/employees"><i class="ion ion-person"></i> Employee</a></li>
-                                <li id="tugboatMenu"><a href="/administrator/maintenance/tugboat"><i class="ion ion-android-boat"></i>Tugboats</a></li>
+                                <li id="positionMenu"><a href="/affiliates/maintenance/position"><i class="fas fa-sitemap"></i>Position</a></li>
+                                <li id="employeesMenu"><a href="/affiliates/maintenance/employees"><i class="ion ion-person"></i> Employee</a></li>
+                                <li id="tugboatMenu"><a href="/affiliates/maintenance/tugboat"><i class="ion ion-android-boat"></i>Tugboats</a></li>
                             @endif
                                 <!--<li id="equipmentsMenu"><a href="/equipment"><i class="fas fa-anchor"></i> Equipment</a></li>-->
                                 {{-- <li id="insurancesMenu"><a href="/administrator/maintenance/insurances"><i class="ion ion-medkit"></i>Insurances</a></li> --}}
@@ -78,10 +77,12 @@
                                         <li id="menuJobOrder"><a class="waves-effect" href="/administrator/transactions/joborders"><i class="ion ion-ios-list"></i>Job Order</a></li>
                                         <li id="menuTugboatAssignment"><a class="waves-effect" href="/administrator/transactions/tugboatassignment"><i class="ion ion-android-boat"></i>Tugboat Assignment</a></li>
                                         <li id="menuTeamBuilder"><a class="waves-effect" href="/administrator/transactions/teamassignment"><i class="ion ion-ios-people"></i>Team Builder</a></li>
+                                        <li id="menuScheduling"><a href="/administrator/transactions/scheduling"><i class="fas fa-calendar-alt"></i>Schedule</a></li>
+                                        <li id="menuHauling"><a href="/administrator/transactions/hauling"><i class="fas fa-link"></i>Hauling</a></li>
+                                        
                                         {{-- <li id="menuScheduling"><a class="waves-effect" href=""><i class="ion ion-ios-time"></i>Scheduling</a></li> --}}
                                     </ul>
                                 </li>
-                                <li id="menuScheduling"><a href="/administrator/transactions/scheduling"><i class="fas fa-calendar-alt"></i>Schedule</a></li>
                                 <li id="menuDispatchTicket"><a href="general.html"><i class="fas fa-ticket-alt"></i>Dispatch Ticket</a></li>
                                 <li id="menuInvoice"><a href="buttons.html"><i class="ion ion-ios-list"></i>Invoice</a></li>
                                 <li id="menuPayment"><a href="buttons.html"><i class="ion ion-cash"></i>Payment &amp; Billing</a></li>