+++ CHANGELOG.md
@@ -1,9 +1,75 @@
-# Changelog Capstone-Laravel-tugboat-dev-PreFinal
-Pre Final Repository Monday August 27,2018
+# Changelog Capstone-Laravel-tugboat-dev-PreFinal α
+### `Pre Final Repository Monday August 27,2018`
 
-# 0.0.4 - 2018-02-29
+&nbsp;
 
-# 0.0.3 - 2018-08-29
+# 0.0.6 - 2018-08-30 α
+
+### added mobile number and telephone number to register forms of the ff:
+
+### Views
+
+- [x] adminRegister
+- [x] affiliatesRegister
+- [x] register
+
+### Routes   
+- [x] AdminLoginController@register
+- [x] AffiliatesLoginController@register
+- [x] UserLoginController@register
+
+# 0.0.5 - 2018-08-29 α
+
+### fixed Middlewares for Admin and Affiliates
+
+- [x] Affiliates are now redirected to `/affiliates/maintenance/employees`
+- [x] Administrator are now redirected to `/administrator/maintenance/pier`
+
+`new`
+### Added Functionality to Affilates
+
+`Maintenance`
+
+- [x] Position
+- [x] Employees
+- [x] Tugboats
+
+##### Routes (Everything needed for Maintenance)
+
+- [x] /position
+- [x] /employees
+- [x] /tugboat
+
+##### Javascript
+
+- [x] positionAffiliates.js
+- [x] employeeAffiliates.js
+- [x] tugboatAffiliates.js
+
+`changes`
+- getdata.js is relocated to /dist/js/tugboat
+- getdata.js is renamed as tugboat.js
+
+# 0.0.4 - 2018-08-29 α
+
+#### fixed the view modal for
+- view initial contract
+- view finalized (active) contract
+
+`additions`
+
+#### user is now notified when 
+
+- the contract is 1 or 2 days from being activated
+- the contract expiration a month before expiry
+- the contract is expired
+
+`fixes`
+#### joborder dropdowns are properly rendering when there is an inactive data
+- Goods
+- Berth
+ 
+# 0.0.3 - 2018-08-29 α
 
 ### Database Changes
 
@@ -24,17 +90,12 @@ Pre Final Repository Monday August 27,2018
 
 ### `new` Routes 
 
-adasdasd
-asdasdasd
-asdasdad
-asdasdasdasd
-
-- [x] Route::post('/berth/activate','BerthController@activate');
-- [x] Route::post('/pier/activate','PierController@activate');
-- [x] Route::post('/position/activate','PositionController@activate');
-- [x] Route::post('/employees/activate','EmployeesController@activate');
-- [x] Route::post('/goods/activate','GoodsController@activate');
-- [x] Route::post('/tugboattype/activate','TugboatTypeController@activate');
+- [x] /berth/activate
+- [x] /pier/activate
+- [x] /position/activate
+- [x] /employees/activate
+- [x] /goods/activate
+- [x] /tugboattype/activate
 
 ### Updated `Controllers` & `Models` 
 
@@ -45,10 +106,9 @@ asdasdasdasd
 - GoodsController::activate -> `Goods`
 - TugboatTypeController::activate -> `TugboatType`
 
-# 0.0.2 - 2018-08-28
+# 0.0.2 - 2018-08-28 α
 All Datatables are set to bg-primary
 
-Added XD
 Added Pills to Consignee Accounts
 - Requests Tab
 - Active Tab
@@ -63,7 +123,7 @@ Bugfixes
 - fixed a bug where the whole tab of Dispacth and Hauling Transaction tends to become active
 
 
-# 0.0.1 - 2018-08-27
+# 0.0.1 - 2018-08-27 α
 Removed Folders and Views
 
 - Agreements